+++ src/main/java/com/ana29/deliverymanagement/security/constant/user/UserRoleEnum.java
@@ -1,4 +1,4 @@
-package com.ana29.deliverymanagement.global.constant;
+package com.ana29.deliverymanagement.security.constant.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
