+++ src/main/java/com/ana29/deliverymanagement/security/jwt/JwtAuthenticationFilter.java
@@ -1,10 +1,10 @@
-package com.ana29.deliverymanagement.security;
+package com.ana29.deliverymanagement.security.jwt;
 
-import com.ana29.deliverymanagement.global.constant.UserRoleEnum;
-import com.ana29.deliverymanagement.global.constant.jwt.JwtConfigEnum;
+import com.ana29.deliverymanagement.security.UserDetailsImpl;
+import com.ana29.deliverymanagement.security.constant.user.UserRoleEnum;
+import com.ana29.deliverymanagement.security.constant.jwt.JwtConfigEnum;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ana29.deliverymanagement.user.dto.SigninRequestDto;
-import com.ana29.deliverymanagement.user.jwt.JwtUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
