+++ src/main/java/com/ana29/deliverymanagement/security/jwt/WebSecurityConfig.java
@@ -1,9 +1,7 @@
-package com.ana29.deliverymanagement.global.config.jwt;
+package com.ana29.deliverymanagement.security.jwt;
 
 
-import com.ana29.deliverymanagement.security.JwtAuthorizationFilter;
 import com.ana29.deliverymanagement.security.UserDetailsServiceImpl;
-import com.ana29.deliverymanagement.user.jwt.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
@@ -18,7 +16,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-import com.ana29.deliverymanagement.security.JwtAuthenticationFilter;
 
 @Configuration
 @EnableWebSecurity // Spring Security 지원을 가능하게 함
