+++ src/test/java/com/ana29/deliverymanagement/mvc/JwtUtilTest.java
@@ -1,8 +1,8 @@
 package com.ana29.deliverymanagement.mvc;
 
-import com.ana29.deliverymanagement.global.constant.UserRoleEnum;
-import com.ana29.deliverymanagement.global.constant.jwt.JwtConfigEnum;
-import com.ana29.deliverymanagement.user.jwt.JwtUtil;
+import com.ana29.deliverymanagement.security.constant.user.UserRoleEnum;
+import com.ana29.deliverymanagement.security.constant.jwt.JwtConfigEnum;
+import com.ana29.deliverymanagement.security.jwt.JwtUtil;
 import io.jsonwebtoken.Claims;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
